requirements.txt
google-cloud-language==2.0.0
google-cloud-speech==2.0.1
moviepy==1.0.3
torch==1.7.1
torchvision==0.8.2
transformers==4.2.0
